chore: add handleFetch function to useReview hook - remain on current sentence when user exceeds the rate limit
